Move PlayLogger to callback plugins default location.. Still haven't been able to activate the plugin.
