+++ subspace/__init__.py
@@ -1,11 +1,11 @@
-from . import constants as _constants
+from . import constants
 from .playbook import PlayBook, get_playbooks
 from .version import VERSION
 import cache
+#TODO: Remove unused imports
 
-
-def constants(key, value=None):
+def set_constants(key, value=None):
     if value is None:
-        return _constants.get(key)
+        return constants.get(key)
     else:
-        _constants.set(key, value)
+        constants.set_value(key, value)
