+++ subspace/playbook_2.py
@@ -1,7 +1,6 @@
 import os
 import operator
 
-from datetime import datetime
 from tempfile import NamedTemporaryFile
 
 from ansible.inventory import Inventory
@@ -9,183 +8,6 @@ from ansible.vars import VariableManager
 from ansible.parsing.dataloader import DataLoader
 from ansible.executor import playbook_executor
 from ansible.utils.display import Display
-from ansible.plugins.callback import CallbackBase
-
-class PlayLogger:
-    """Store log output in a single object.
-    We create a new object per Ansible run
-    """
-    def __init__(self):
-        self.log = ''
-        self.runtime = 0
-
-    def append(self, log_line):
-        """append to log"""
-        self.log += log_line+"\n\n"
-
-    def banner(self, msg):
-        """Output Trailing Stars"""
-        width = 78 - len(msg)
-        if width < 3:
-            width = 3
-        filler = "*" * width
-        return "\n%s %s " % (msg, filler)
-
-
-class CallbackModule(CallbackBase):
-    """
-    Reference: https://github.com/ansible/ansible/blob/v2.0.0.2-1/lib/ansible/plugins/callback/default.py
-    """
-
-    CALLBACK_VERSION = 2.0
-    CALLBACK_TYPE = 'stored'
-    CALLBACK_NAME = 'database'
-
-    def __init__(self):
-        super(CallbackModule, self).__init__()
-        self.logger = PlayLogger()
-        self.start_time = datetime.now()
-
-    def v2_runner_on_failed(self, result, ignore_errors=False):
-        delegated_vars = result._result.get('_ansible_delegated_vars', None)
-
-        # Catch an exception
-        # This may never be called because default handler deletes
-        # the exception, since Ansible thinks it knows better
-        if 'exception' in result._result:
-            # Extract the error message and log it
-            error = result._result['exception'].strip().split('\n')[-1]
-            self.logger.append(error)
-
-            # Remove the exception from the result so it's not shown every time
-            del result._result['exception']
-
-        # Else log the reason for the failure
-        if result._task.loop and 'results' in result._result:
-            self._process_items(result)  # item_on_failed, item_on_skipped, item_on_ok
-        else:
-            if delegated_vars:
-                self.logger.append("fatal: [%s -> %s]: FAILED! => %s" % (result._host.get_name(), delegated_vars['ansible_host'], self._dump_results(result._result)))
-            else:
-                self.logger.append("fatal: [%s]: FAILED! => %s" % (result._host.get_name(), self._dump_results(result._result)))
-
-    def v2_runner_on_ok(self, result):
-        self._clean_results(result._result, result._task.action)
-        delegated_vars = result._result.get('_ansible_delegated_vars', None)
-        if result._task.action == 'include':
-            return
-        elif result._result.get('changed', False):
-            if delegated_vars:
-                msg = "changed: [%s -> %s]" % (result._host.get_name(), delegated_vars['ansible_host'])
-            else:
-                msg = "changed: [%s]" % result._host.get_name()
-        else:
-            if delegated_vars:
-                msg = "ok: [%s -> %s]" % (result._host.get_name(), delegated_vars['ansible_host'])
-            else:
-                msg = "ok: [%s]" % result._host.get_name()
-
-        if result._task.loop and 'results' in result._result:
-            self._process_items(result)  # item_on_failed, item_on_skipped, item_on_ok
-        else:
-            self.logger.append(msg)
-
-    def v2_runner_on_skipped(self, result):
-        if result._task.loop and 'results' in result._result:
-            self._process_items(result)  # item_on_failed, item_on_skipped, item_on_ok
-        else:
-            msg = "skipping: [%s]" % result._host.get_name()
-            self.logger.append(msg)
-
-    def v2_runner_on_unreachable(self, result):
-        delegated_vars = result._result.get('_ansible_delegated_vars', None)
-        if delegated_vars:
-            self.logger.append("fatal: [%s -> %s]: UNREACHABLE! => %s" % (result._host.get_name(), delegated_vars['ansible_host'], self._dump_results(result._result)))
-        else:
-            self.logger.append("fatal: [%s]: UNREACHABLE! => %s" % (result._host.get_name(), self._dump_results(result._result)))
-
-    def v2_runner_on_no_hosts(self, task):
-        self.logger.append("skipping: no hosts matched")
-
-    def v2_playbook_on_task_start(self, task, is_conditional):
-        self.logger.append("TASK [%s]" % task.get_name().strip())
-
-    def v2_playbook_on_play_start(self, play):
-        name = play.get_name().strip()
-        if not name:
-            msg = "PLAY"
-        else:
-            msg = "PLAY [%s]" % name
-
-        self.logger.append(msg)
-
-    def v2_playbook_item_on_ok(self, result):
-        delegated_vars = result._result.get('_ansible_delegated_vars', None)
-        if result._task.action == 'include':
-            return
-        elif result._result.get('changed', False):
-            if delegated_vars:
-                msg = "changed: [%s -> %s]" % (result._host.get_name(), delegated_vars['ansible_host'])
-            else:
-                msg = "changed: [%s]" % result._host.get_name()
-        else:
-            if delegated_vars:
-                msg = "ok: [%s -> %s]" % (result._host.get_name(), delegated_vars['ansible_host'])
-            else:
-                msg = "ok: [%s]" % result._host.get_name()
-
-        msg += " => (item=%s)" % (result._result['item'])
-
-        self.logger.append(msg)
-
-    def v2_playbook_item_on_failed(self, result):
-        delegated_vars = result._result.get('_ansible_delegated_vars', None)
-        if 'exception' in result._result:
-            # Extract the error message and log it
-            error = result._result['exception'].strip().split('\n')[-1]
-            self.logger.append(error)
-
-            # Remove the exception from the result so it's not shown every time
-            del result._result['exception']
-
-        if delegated_vars:
-            self.logger.append("failed: [%s -> %s] => (item=%s) => %s" % (result._host.get_name(), delegated_vars['ansible_host'], result._result['item'], self._dump_results(result._result)))
-        else:
-            self.logger.append("failed: [%s] => (item=%s) => %s" % (result._host.get_name(), result._result['item'], self._dump_results(result._result)))
-
-    def v2_playbook_item_on_skipped(self, result):
-        msg = "skipping: [%s] => (item=%s) " % (result._host.get_name(), result._result['item'])
-        self.logger.append(msg)
-
-    def v2_playbook_on_stats(self, stats):
-        run_time = datetime.now() - self.start_time
-        self.logger.runtime = run_time.seconds  # returns an int, unlike run_time.total_seconds()
-
-        hosts = sorted(stats.processed.keys())
-        for h in hosts:
-            t = stats.summarize(h)
-
-            msg = "PLAY RECAP [%s] : %s %s %s %s %s" % (
-                h,
-                "ok: %s" % (t['ok']),
-                "changed: %s" % (t['changed']),
-                "unreachable: %s" % (t['unreachable']),
-                "skipped: %s" % (t['skipped']),
-                "failed: %s" % (t['failures']),
-            )
-
-            self.logger.append(msg)
-
-    def record_logs(self, user_id, success=False):
-        """
-        Special callback added to this callback plugin
-        Called by Runner objet
-        :param user_id:
-        :return:
-        """
-        print "When this works i will 'store' the log results somewhere."
-        print "For now: User: %s Runtime: %s - All logs:\n%s" % (user_id, self.logger.runtime, self.logger.log)
-
 
 
 class Options(object):
@@ -198,7 +20,8 @@ class Options(object):
                  sudo=None, sudo_user=None, become=None, become_method=None, become_user=None, become_ask_pass=None,
                  ask_pass=None, private_key_file=None, remote_user=None, connection=None, timeout=None, ssh_common_args=None,
                  sftp_extra_args=None, scp_extra_args=None, ssh_extra_args=None, poll_interval=None, seconds=None, check=None,
-                 syntax=None, diff=None, force_handlers=None, flush_cache=None, listtasks=None, listtags=None, module_path=None):
+                 syntax=None, diff=None, force_handlers=None, flush_cache=None, listtasks=None, listtags=None, module_path=None,
+                 logger=None):
         self.verbosity = verbosity
         self.inventory = inventory
         self.listhosts = listhosts
@@ -241,12 +64,13 @@ class Options(object):
         self.listtasks = listtasks
         self.listtags = listtags
         self.module_path = module_path
+        self.logger = logger
 
 
 class Runner(object):
 
     def __init__(self, hosts_file, playbook, private_key_file, run_data,
-                 limit_hosts=None, group_vars_map={}, become_pass=None, verbosity=4):
+                 limit_hosts=None, group_vars_map={}, become_pass=None, verbosity=0, logger=None):
 
         self.run_data = run_data
 
@@ -255,6 +79,7 @@ class Runner(object):
                 private_key_file=private_key_file,
                 subset=limit_hosts,
                 connection='ssh',  # Need a connection type "smart" or "ssh"
+                logger=logger,
                 become=True,
                 become_method='sudo',
                 become_user='root',
@@ -296,6 +121,7 @@ class Runner(object):
         if self.options.subset:
             hostname = self.options.subset['ip']  # ['hostname']
             self.inventory.subset(hostname)
+
         # Playbook to run. Assumes it is
         # local to this python file
         if isinstance(playbook, basestring):
@@ -322,6 +148,7 @@ class Runner(object):
                 if os.path.exists(file_path):
                     self.variable_manager.add_group_vars_file(file_path,self.loader)
             print "Vars found: %s" % variables
+
         # Setup playbook executor, but don't run until run() called
         self.pbex = playbook_executor.PlaybookExecutor(
             playbooks=playbooks,
@@ -330,6 +157,11 @@ class Runner(object):
             loader=self.loader,
             options=self.options,
             passwords=None)
+
+        self.pbex._tqm.send_callback(
+            'set_logger',
+            logger=self.options.logger
+        )
 
     def run(self):
         # Results of PlaybookExecutor in stats.
@@ -393,6 +225,7 @@ def get_playbook_runner(playbook, **kwargs):
     else:
         run_data = kwargs.get('run_data', {})
     host_file = kwargs.get('host_list')
+    logger = kwargs.get('logger')
     group_vars_map = kwargs.get('group_vars_map')
     private_key_file = kwargs.get('private_key', 'No Private Key provided')
     print "Use key: %s" % private_key_file
@@ -403,7 +236,8 @@ def get_playbook_runner(playbook, **kwargs):
         run_data=run_data,
         group_vars_map=group_vars_map,
         limit_hosts=kwargs.get('limit_hosts'),
-        verbosity=4
+        verbosity=4,
+        logger=logger
     )
     return runner
